+++ packages/web/src/index.ts
@@ -8,6 +8,7 @@ export {
   Extension,
   MagicSDKError as SDKError,
   MagicExtensionError as ExtensionError,
+  MagicExtensionWarning as ExtensionWarning,
   MagicRPCError as RPCError,
   MagicSDKWarning as SDKWarning,
   MagicSDKAdditionalConfiguration,